backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
transformers>=4.36.0
python-dotenv>=1.0.0
peft>=0.7.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
drf-yasg>=1.21.7  # For API documentation
django-filter>=23.5.0
Pillow>=10.1.0  # For image handling
python-jose>=3.3.0  # For JWT handling
requests>=2.31.0
